+++ src/tables/const_column.hpp
@@ -13,12 +13,12 @@ namespace tables
         /// Creates a constant column
         /// @param value The value stored by this column
         /// @param rows The initial number of rows
-        const_column(const boost::any& value, uint32_t rows = 0);
+        const_column(const boost::any& value, uint64_t rows = 0);
 
     public:
 
-        /// @copydoc column::value(uint32_t) const
-        boost::any value(uint32_t row_index) const;
+        /// @copydoc column::value(uint64_t) const
+        boost::any value(uint64_t row_index) const;
 
         /// @copydoc column::default_value() const
         boost::any default_value() const;
@@ -26,8 +26,8 @@ namespace tables
         /// @copydoc column::values() const
         std::vector<boost::any> values() const;
 
-        /// @copydoc column::add_rows(uint32_t)
-        void add_rows(uint32_t rows);
+        /// @copydoc column::add_rows(uint64_t)
+        void add_rows(uint64_t rows);
 
         /// @copydoc column::add_row()
         void add_row();
@@ -56,6 +56,6 @@ namespace tables
         const boost::any m_value;
 
         /// Keeps track of the column's size
-        uint32_t m_rows;
+        uint64_t m_rows;
     };
 }
